[Refactor]CampusServiceV2 클래스 분리
